Use slack alert color function
