+++ src/sentry/integrations/message_builder.py
@@ -5,9 +5,11 @@ from typing import Any, Callable, Mapping, Sequence
 
 from sentry import features
 from sentry.eventstore.models import GroupEvent
-from sentry.integrations.slack.message_builder import SLACK_URL_FORMAT
+from sentry.integrations.slack.message_builder import LEVEL_TO_COLOR, SLACK_URL_FORMAT
+from sentry.issues.grouptype import GroupCategory
 from sentry.models import Group, Project, Rule, Team
 from sentry.notifications.notifications.base import BaseNotification
+from sentry.notifications.notifications.rules import AlertRuleNotification
 from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.types.integrations import EXTERNAL_PROVIDERS, ExternalProviders
 from sentry.utils.dates import to_timestamp
@@ -154,3 +156,27 @@ def build_footer(
 def get_timestamp(group: Group, event: GroupEvent | None) -> float:
     ts = group.last_seen
     return to_timestamp(max(ts, event.datetime) if event else ts)
+
+
+def get_color(
+    event_for_tags: GroupEvent | None, notification: BaseNotification | None, group: Group
+) -> str:
+    if notification:
+        if not isinstance(notification, AlertRuleNotification):
+            return "info"
+    if event_for_tags:
+        color: str | None = event_for_tags.get_tag("level")
+        if (
+            hasattr(event_for_tags, "occurrence")
+            and event_for_tags.occurrence is not None
+            and event_for_tags.occurrence.level is not None
+        ):
+            color = event_for_tags.occurrence.level
+        if color and color in LEVEL_TO_COLOR.keys():
+            return color
+    if group.issue_category == GroupCategory.PERFORMANCE:
+        # XXX(CEO): this shouldn't be needed long term, but due to a race condition
+        # the group's latest event is not found and we end up with no event_for_tags here for perf issues
+        return "info"
+
+    return "error"
