+++ src/sentry/integrations/slack/message_builder/issues.py
@@ -10,16 +10,15 @@ from sentry.integrations.message_builder import (
     build_attachment_title,
     build_footer,
     format_actor_options,
+    get_color,
     get_timestamp,
     get_title_link,
 )
-from sentry.integrations.slack.message_builder import LEVEL_TO_COLOR, SLACK_URL_FORMAT, SlackBody
+from sentry.integrations.slack.message_builder import SLACK_URL_FORMAT, SlackBody
 from sentry.integrations.slack.message_builder.base.base import SlackMessageBuilder
 from sentry.integrations.slack.utils.escape import escape_slack_text
-from sentry.issues.grouptype import GroupCategory
 from sentry.models import ActorTuple, Group, GroupStatus, Project, Rule, Team, User
-from sentry.notifications.notifications.base import BaseNotification, ProjectNotification
-from sentry.notifications.notifications.rules import AlertRuleNotification
+from sentry.notifications.notifications.base import ProjectNotification
 from sentry.notifications.utils.actions import MessageAction
 from sentry.services.hybrid_cloud.actor import ActorType, RpcActor
 from sentry.services.hybrid_cloud.identity import RpcIdentity, identity_service
@@ -192,30 +191,6 @@ def build_actions(
     )
 
     return [resolve_button, ignore_button, assign_button], text, color
-
-
-def get_color(
-    event_for_tags: GroupEvent | None, notification: BaseNotification | None, group: Group
-) -> str:
-    if notification:
-        if not isinstance(notification, AlertRuleNotification):
-            return "info"
-    if event_for_tags:
-        color: str | None = event_for_tags.get_tag("level")
-        if (
-            hasattr(event_for_tags, "occurrence")
-            and event_for_tags.occurrence is not None
-            and event_for_tags.occurrence.level is not None
-        ):
-            color = event_for_tags.occurrence.level
-        if color and color in LEVEL_TO_COLOR.keys():
-            return color
-    if group.issue_category == GroupCategory.PERFORMANCE:
-        # XXX(CEO): this shouldn't be needed long term, but due to a race condition
-        # the group's latest event is not found and we end up with no event_for_tags here for perf issues
-        return "info"
-
-    return "error"
 
 
 class SlackIssuesMessageBuilder(SlackMessageBuilder):
